renaming things & fixing file explorer automatic preview bug
