+++ src/dawg/extensions/core/controls/VerticalSwitch.vue
