+++ src/dawg/extensions/core/patterns/index.ts
@@ -1,60 +1,49 @@
 import Vue from 'vue';
 import Patterns from '@/dawg/extensions/core/patterns/Patterns.vue';
-import { ref, computed, createComponent } from '@vue/composition-api';
+import { ref, computed, createComponent, watch } from '@vue/composition-api';
 import { makeLookup, vueExtend } from '@/utils';
 import { Pattern } from '@/core';
 import { manager } from '@/base/manager';
 import { ui } from '@/base/ui';
 import { project } from '@/dawg/extensions/core/project';
+import * as t from '@/modules/io';
 
 export const patterns = manager.activate({
   id: 'dawg.patterns',
-  activate() {
-    const selectedPatternId = ref<null | string>(null);
+  workspace: {
+    selectedPatternId: t.string,
+  },
+  activate(context) {
+    const selectedPatternId = context.workspace.selectedPatternId;
+    const pattern = ref<Pattern>();
 
     const patternLookup = computed(() => {
       return makeLookup(project.patterns);
     });
 
-    const selectedPattern = computed(() => {
-      if (selectedPatternId.value === null) { return null; }
-      const p = patternLookup.value[selectedPatternId.value];
-      if (!p) {
-        return null;
+    if (selectedPatternId.value) {
+      pattern.value = patternLookup.value[selectedPatternId.value];
+    }
+
+    watch(pattern, () => {
+      selectedPatternId.value = pattern.value ? pattern.value.id : undefined;
+
+      if (pattern.value && ui.openedSideTab.value !== 'Patterns') {
+        ui.openedSideTab.value = 'Patterns';
       }
-
-      return p;
-    });
-
-    const setPattern = (pattern: Pattern | null) => {
-      if (pattern) {
-        selectedPatternId.value = pattern.id;
-      } else {
-        selectedPatternId.value = null;
-      }
-    };
-
-    Vue.extend({
-      props: [],
-    });
-
-    Vue.extend({
-      props: {},
     });
 
     const wrapper = vueExtend(createComponent({
       components: { Patterns },
       template: `
       <patterns
-        :value="selectedPattern"
+        v-model="pattern"
         :patterns="patterns"
-        @input="setPattern"
         @remove="remove"
       ></patterns>
       `,
       setup: () => ({
-        selectedPattern,
-        setPattern,
+        pattern,
         patterns: project.patterns,
         remove: (i: number) => project.removePattern(i),
       }),
@@ -74,7 +63,7 @@ export const patterns = manager.activate({
     });
 
     return {
-      selectedPattern,
+      selectedPattern: pattern,
     };
   },
 });
