+++ src/dawg/index.ts
@@ -9,7 +9,7 @@ export { instruments } from '@/dawg/extensions/core/instruments';
 export { ui } from '@/base/ui';
 export { pianoRoll } from '@/dawg/extensions/core/piano-roll';
 export { patterns } from '@/dawg/extensions/core/patterns';
-export { applicationContext } from '@/dawg/extensions/core/application-context';
+export { controls } from '@/dawg/extensions/core/controls';
 export { menubar } from '@/dawg/extensions/core/menubar';
 export { record } from '@/dawg/extensions/core/record';
 export { status } from '@/dawg/extensions/core/status';
